Switch to using ts-node to run the prod app
Because it's better to keep local dev and prod as close to eachother as possible to reduce the chance of prod bugs that can't be reproduced locally.

Per https://github.com/TypeStrong/ts-node/issues/104 it seems the main impact of using it in production is a slightly increased memory footprint.